Update Routes
I took an initial pass at some routes. I introduced the concept of "Players".
This would be a players profile across multiple games. I also added "Games"
which contain many "Rounds" & "Players". I'm not to clear on how you envisioned
this working, but I was thinking that a game could be initialized and then
players would be added to the game, then rounds would be added until the game
finished.