Wombats Update
This PR includes work that was done outside of the Wombats project
as an effort to learn more about how to create a web service in
Clojure. This PR takes that work to create a foundation for moving
forward with Wombats.

Changes include:

- Clojure 1.8.x -> Clojure 1.9.0alpha-14
- Added component lib
- schema -> core.spec
- ring -> pedestal.io
- ring handlers / ring middleware -> pedeatal interceptors / async handlers
- mongodb -> datomic

All other code has been moved under wombats-api-legacy to be ported over.