Mock AWS Lambda for development.
This is triggered by leaving out/empty the AWS credentials or Lambda
settings in the `~/.wombats/config.edn` file.

For Python and JavaScript, this works by creating a temporary file with
the wombats-bot function and some boilerplate to call it on JSON from
STDIN and write the output as JSON to STDOUT.

Fixes #329.